Preliminary implementation of functional BST
**This Commit**
Adds the beginnings of a functional-style BST taking inspiration from
Chris's SuperMegaSexyHaskellBST.

**Why?**
Because I'm having a really hard time with the non-functional ones and
Chris already slayed this one so it seemed like easy pickin's.

**Todo**
This commit only introduces methods for inserting and finding. We'll
need deletion later and much better tests (and iterators and all the
good stuff!).