Add criterion benchmarks
**This Commit**
Adds some criterion benchmarks for the functional BST.

**Why?**
So we can assess the performance impact of balancing the tree and
micro-optimizations.

**Note**
These benchmarks intentionally only test what happens when we mess with
the rightmost node. This is because, until we add self-balancing,
messing with other nodes (notably the root node) can have super
surprising results (because an unbalanced tree might not have a
predecessor that the balanced tree has, etc.).

We'll start with these tests just so we can add self-balancing and show
that the performance is unaffected by how the values are inserted. Then
we can add benchmarks for operations on the root node and test further.