Add rotations to functional AVL tree.
**Why?**
To make this a proper self-balancing AVL tree!

**Note**
Some unit tests are now sketchy in that they were intending to set up a
particular tree but, with balancing happening, the resulting tree may
not match. A future PR should create a macro that allows you to declare
the tree and it just gets built without calls to `insert`.