Implement Tree methods on Node
## This Commit

Moves the implementation of `insert`, `delete`, and `find` from the
`Tree` struct to the `Node` struct.

## Why?

In the next PR we're going to change `Tree` from an enum to an
`Option<Node>`. When we do this, we'll start calling things like
`self.0.insert`. So we'll want to move the methods onto `Node`. By doing
so now, we reduce the diff later and we can more accurately assess any
performance changes.