Add double rotations for functional BST
## This Commit

Updates the balancing logic to handle the case where a double rotation
is needed. It also adjusts a `cfg!(test)` that isn't as active as it
could be.

## Why?

I was writing the `unsafe` tree and noticed that I didn't handle double
rotations. Then I remembered that I didn't write them for the functional
tree either. I noticed the `cfg!(test)` assertion wasn't active during
quickchecks because they're run in integration tests which compile the
library as non-`test`. I updated it to `cfg!(debug_assertions)` and they
caught this bug.