Move quickchecks out of integration tests
## Why?

Mostly just for expectations. I like the idea of fuzzing the public
interface. That's still happening. But I also like being able to fuzz
the private interface if need be. Also, my `cfg!(test)` mistake before
was because I expected the quickchecks to validate this behavior and
they didn't because they were integration tests. So this should be more
clear without sacrificing anything I'm aware of.