Rename `functional` module to `immutable`
## Why?

I think this is probably more accurate/descriptive. It's not like this
BST has functions as first class citizens. I mean, the API is fluent
(i.e. `insert`/`delete` return `Self`) but that's probably not exactly
what people are looking for here. The real point is that we don't mutate
the tree. So I think this is more correct.