Extract `Link` type for unsafe rotations
## This Commit

Extracts a `Link` type to implement on rotations so root and non-root
nodes can rotate purely through pointer swapping.

## Why?

This method of rotation is simpler to understand and maintain compared
to the previous `mem::swap` approach. Also, for nodes with large
keys/values, moving pointers will be faster than `mem::swap`ping node
fields.

## Benchmarks

With the change of benchmarks to have nodes with `[i32; 1024]`, it seems
to make an improvement but it didn't help for nodes with `i32` values.
We might want a constant time check to decide between the two based on
the types of `K` and `V`.