Add tree-based iterator to immutable tree
## This Commit

Allows for iterating over references of the key-value pairs in an
immutable tree.

## Why?

Because this is a standard collections interface we should support. Also
it's a good learning opportunity because it's sneaky-hard.

## Up Next

I do not like what I came up with one bit haha. I know I saw something
about this in the Nomicon way back when so I'm going to go look that up
now and update with the changes.

### Update

Well, after looking at the Nomicon, it seems like this is more-or-less
The Way so :tada:?